Migrated API functions to external file. Updated spring-boot cross-origin to accept https domains
